fix the punctuation characters
